requirements.txt
####### BASE #######
numpy
spotipy
tqdm
lyricsgenius
bs4
seaborn
####################
###  PYMAGNITUDE ###
pymagnitude
lz4
xxhash
annoy
fasteners
# https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
torchvision_nightly
h5py
####################

####### DEV ########
pylint
autopep8
rope
####################

###### CLASS #######
sentencepiece
gensim
####################
